backend/pyproject.toml
[project]
name = "agents-backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.68.0",
  "uvicorn[standard]",
  "python-multipart",
  "pydantic[email]",
  "openai==1.86.0",
  "openai-agents>=0.0.17", # OpenAI Agents SDK for multi-agent workflows
  "sqlalchemy>=2.0.0",
  "psycopg2-binary",
  "alembic",
  "python-dotenv",
  "ag-ui-protocol>=0.1.5",
  "websockets>=11.0",
  "python-multipart>=0.0.5",
  # CV Extraction Dependencies (Phase 1-3)
  "pgvector>=0.2.0", # PgVector Python client
  "pypdf2>=3.0.0", # PDF text extraction (fallback)
  "python-docx>=0.8.11", # Word document extraction
  "tiktoken>=0.5.0", # Token counting for embeddings
  "mistralai>=1.0.0", # Mistral OCR API client (future use)
  "pillow>=10.0.0", # Image processing for vision models
  "crewai>=0.130.0",
  "crewai-tools>=0.47.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
]
test = [
  "pytest>=7.4.0",
  "pytest-asyncio>=0.21.0",
  "httpx>=0.24.0",
  "pytest-cov>=4.1.0",
  "pytest-mock>=3.11.0",
  "pytest-timeout>=0.4.2",
]

[project.scripts]
test = "pytest"
test-cov = "pytest --cov=. --cov-report=html --cov-report=term"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --durations=10"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
# Hard per-test timeout (seconds)
timeout = 15

[tool.coverage.run]
source = ["."]
omit = [
  "tests/*",
  ".venv/*",
  "__pycache__/*",
  "*.pyc",
]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "raise AssertionError",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
]
